Apt: XB1-audit corrections for the real native-8 GUIAPT bundles
The Xbox One retail i64 (the only 64-bit build with Apt symbols; exported
to .ida-exports/Burnout_External_Xbox_One.exe, 77107 fns) is now the
x64-ABI arbiter. Audit findings applied:

- ZID chain widened to pointer width (XB1 ?GetZID@AptCharacterTextInst
  does an 8-byte load @item+0x68): AptRenderItemDynamicText::mZID +
  CustomControl::mZId are intptr_t handles end-to-end (facade, hooks,
  sentinel, use sites). The dynamic-text member block shifts +4 after it
  (XB1 offsets documented in the header).
- tag-4 remove / tag-5 back-to-script payloads read @cmd+8 (XB1 asm;
  EMPIRICALLY confirmed vs B5MENUITEM.bundle -- cmd+4 is the tag pad, so
  on the real drive bundles every timeline remove had silently no-op'd
  again).
- resolve64: clipActions block recArrayPtr @block+0x08 (XB1/native-8);
  label-hash init zeroes the four 8-BYTE member slots (the 4-byte zeroing
  left x64 pointer halves as pool garbage). Converter-delta comments
  retired with the converter (the staged GUIAPT set is now JeBobs' real
  native-8 drop; the .apt4 intermediates are deleted).
- DoTemporaryFrameControls merge-onto-existing rewritten against the
  typed x64 AptPseudoData_t per XB1 sub_1408363A0: store the ADDRESSES of
  the record's matrix/colour fields + the ratio value + OR the flags. The
  old console transliteration read a half-pointer at node+4 and wrote
  4-byte values through it -- wild heap writes on every MOVE-over-existing
  merge (prime suspect for the recurring 0xc0000374).
- Unresolved-import guards on the seek/replay path (drive bundles keep
  imports as REAL import-table entries): DoTempFC skips a fresh place
  whose charTable slot is null, and AddToDisplayList skips a fresh add
  whose placement names no character (both mirrored from the live place
  path; the null-character instantiate AV'd on menu state jumps).
- tick end-wrap: the faithful jumpToFrame(0) was tried and reproducibly
  killed composition on the real bundles -- kept as the plain play-head
  reset with the diagnosis FLAG'd (suspects: replay over unresolved-import
  clips / the delay-release chain).
- PC keyboard poll is focus-gated (GetAsyncKeyState is global; terminal
  keystrokes were driving the boot flow).
- AptAnimationTarget +0x38 named mpTopMostSprite (XB1 GetTopMostSprite).

Boot-verified on the drive bundles: title composes, menu opens, full
up/down state-jump stress survives. KNOWN GAP (one root cause, three
symptoms, next milestone): import-EXPORT resolution -- menu item/prompt
text ("no dynamic-text field in subtree"), HDComp_mc 'transin' label,
and the ControllerButtons glyph all wait on Fixup pass-3 resolving
imports to their exported characters.

+++ src/GameSource/Gui/BrnAptRuntimeBringUp.cpp
@@ -201,22 +201,22 @@ extern AptActionInterpreter gAptActionInterpreter;
 // callbacks directly; here the bring-up installs bridges that forward to CgsGui::AptCallbackRender.
 // The engine passes the ZID (AptRenderItemDynamicText::mZID) -- our x64 slot-index handle.
 enum AptMaskRenderOperation : int;
-extern void (*gpfnAptDrawTextRenderData)(int nZId, AptMaskRenderOperation eOp, int nTick);  // dword_8324E868
-extern void (*gpfnAptReleaseTextRenderData)(int nZId, int nOp);                             // dword_8324E864
+extern void (*gpfnAptDrawTextRenderData)(intptr_t nZId, AptMaskRenderOperation eOp, int nTick);  // dword_8324E868
+extern void (*gpfnAptReleaseTextRenderData)(intptr_t nZId, int nOp);                        // dword_8324E864
 
 // The bring-up bridges the two hooks forward to. DrawString batches the ZID's glyphs;
 // DeallocateString frees the ZID's pooled string. (leFlags/nOp carried for parity.)
-static void AptRT_DrawTextRenderData(int nZId, AptMaskRenderOperation eOp, int nTick)
+static void AptRT_DrawTextRenderData(intptr_t nZId, AptMaskRenderOperation eOp, int nTick)
 {
     (void)nTick;   // the render callback keys off the ZID + mask op (level unused on this path)
     CgsGui::AptCallbackRender::DrawString(
-        reinterpret_cast<AptAssetString>(static_cast<intptr_t>(nZId)), eOp, 0);
+        reinterpret_cast<AptAssetString>(nZId), eOp, 0);
 }
 
-static void AptRT_ReleaseTextRenderData(int nZId, int nOp)
+static void AptRT_ReleaseTextRenderData(intptr_t nZId, int nOp)
 {
     CgsGui::AptCallbackRender::DeallocateString(
-        reinterpret_cast<AptAssetString>(static_cast<intptr_t>(nZId)),
+        reinterpret_cast<AptAssetString>(nZId),
         static_cast<u32>(nOp));
 }
